+++ somni-expr/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Removed `Context::evaluate_expr_any` and `evaluate_any`. You can specify `evaluate` (and `evaluate_parsed`) to return `TypedValue`.
 - Merged the `Load` and `Store` traits into `LoadStore`.
 - Added `ExpressionVisitor::visit_right_hand_expression`.
+- Added `TypedValue::modulo` for modulo operation.
 
 # [0.1.0] - 2025-07-24